+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Update donation dialog breakpoints
+
 ## 2.2.2 (2019-05-09)
 
 * Add MUI Radio components from Material UI to list of exported MUI components